requirements.txt
Django==5.1.5
openpyxl==3.1.5
gunicorn==23.0.0
python-dotenv==1.0.1
